+++ NEWS.md
@@ -1,6 +1,8 @@
 ## Unreleased changes
 - Fixed compatibility with LLVM 17.
 - Fixed an intermittent crash when using the `--jit` elaboration option.
+- Improved overload resolution when subprogram argument is known to have
+  character type.
 
 ## Version 1.10.3 - 2023-09-17
 - Fixed memory corruption in rare circumstances with functions declared